SitePulse/backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
sqlalchemy==2.0.35
requests==2.32.3
python-dotenv==1.0.1
pydantic
psycopg2-binary
